+++ libs/help/help/src/app/public/modules/shared/initialization.service.ts
@@ -1,5 +1,6 @@
 import {
-  Injectable
+  Injectable,
+  Optional
 } from '@angular/core';
 
 import {
@@ -18,14 +19,24 @@ import {
   HelpWidgetConfig
 } from './widget-config';
 
+import {
+  InitializationConfigExtensionService
+} from './initialization-config-extension.service';
+
+import {
+  Observable,
+  of
+} from 'rxjs';
+
 @Injectable()
 export class HelpInitializationService {
   public constructor(
     private windowRef: SkyAppWindowRef,
-    private config: SkyAppConfig
+    private config: SkyAppConfig,
+    @Optional() private configExtensionService: InitializationConfigExtensionService = undefined
   ) { }
 
-  public load(config: HelpWidgetConfig = {}) {
+  public load(config: HelpWidgetConfig = {}): Promise<void> {
     if (this.config.runtime.params.has('svcid')) {
       config.extends = this.config.runtime.params.get('svcid');
     }
@@ -39,6 +50,10 @@ export class HelpInitializationService {
       const languages = skyuxHost.acceptLanguage || '';
       config.locale = languages.split(',')[0];
     }
-    return BBHelpClient.load(config);
+    const configObservable: Observable<HelpWidgetConfig> = this.configExtensionService
+      ? this.configExtensionService.extend(config)
+      : of(config);
+    return configObservable.toPromise()
+      .then((extendedConfig: HelpWidgetConfig) => BBHelpClient.load(extendedConfig));
   }
 }